DO-F25a: don't disable NM, disable dns for NM
We finally got some fixes in to F25 for NM. The only thing we
have to do now is disable NM from touching /etc/resolv.conf. DO
sets that via cloud-init and we don't want NM stomping on it.

The only reason NM stomps on it now is because when the instance
first boots it is configured to bring up eth0 via DHCP. Once the
static network configuration kicks in NM would be good and not
touch /etc/resolv.conf, but it's too late at that point.